requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv
requests
beautifulsoup4
PyPDF2
tabula-py
pandas